Handle search results being cleared in `componentWillUnmount`
Usually for universally rendered components we rely on `fetchData` to
return the results we need. `fetchData` is called when the component is
created which is usually equivalent to `componentWillMount`. However in
this case, the Search component clears search results in
`componentWillUnmount`. Due to how react life cycle methods work,
`fetchData` is called before the previous components
`componentWillUnmount`.

This means that each time we render the Search container, we're
triggering two search actions.

On the client:

- fetchData fires a search action but thinks it has cached results
- componentWillMount gets called after the results have been cleared, so
  it fetches results

On the server:

- fetchData fetches the results
- componentWillMount fires the search action but has cached results

Until we come up with a better place to call `fetchData` this is going
to be necessary.