Introduce new `auth` and `device` context
This means that:
- any component can get access to the objects related to the
  authenticated user.
- any component can get access to responsive attributes (ie.
  largerDevice).

Previously, we only had access to these if we connected the components,
so we ended up passing down `largerDevice` to multiple levels. Now if a
component depends on those values, they can grab them from the context.

NOTE: not everything should go into the context, auth and responsive
state are pervasive enough where it makes sense.