Support specifying required fields for normalized protobufs.
We had an issue here where `permissions` wasn't inflated on the PostV1
object that is returned with `get_posts`. When you went to edit the
draft, it had the post without permissions in the cache, and rendered
that one. Once the fully inflated post had been fetched, it re-rendered
with the publish button.

Now we wait for the fully inflated post before displaying the draft.