Add express-session to support redis backed sessions.
We need this for handling SSO. When we receive the response from the SSO
provider, we call `completeAuthorization` and then `authenticate` with
the details from `completeAuthorization`. The current issue is that the
`authenticate` call returns a response header of `set-cookie` with the
auth token. On the client side, the browser sets the cookie so when we
redirect to `/` or some other authenticated route, the cookie is present
and we return the logged in information. With server side rendering, all
of these requests get handled by the server and the cookie is never set
in the subsequent requests.

Using the redis session, before redirecting to the authenticated route
after calling `authenticate` on the server, we can store the auth token
in the session and reload it on the next request from the user.