Specify content as a required field for displaying a post
getPosts excludes fetching the content. Previously, if we clicked
through to a post from getPosts, the cache would return a post without
content. Since the Post component expects content to be present, we
would get an error until the full post came in from the `fetchPost` call
for the Post container.

Now we specify "content" as a required field when fetching a post from
the cache, so if content isn't present, we'll wait for the full object
to come in.