Add a CLIENT_MOUNTED action.
Components that must be rendered on the client only (ie. infinite scroll
lists) can now check the `device.mounted` context to determine whether
or not the client is present.

We can't simply check the __CLIENT__ global because react will complain
that the server rendered markup doesn't match the client markup. This
allows the server and initial client render to be the same, and we can
then transition in client dependent components.