Support pagination with cached objects.
Previously, if we loaded a bunch of content for an infinite scroll list
into the cache, left and came back to the component (ie. switching from
"People" and "About" tabs on TeamDetail) it would take forever to load
the page because we would be rendering all of the cached items in the
infinite scroll list.

Now with the bailout action, we can more efficiently slice the data we
want to show based on what page we should be on from the nextRequest.
Even if the nextRequest bails because we have content in the cache, we
can increment the `currentPage` in the paginator reducer and slice the
correct new amount of data.