Fix issue where we couldn't leave the FormPeopleSelector
There was an issue where once focused the FormPeopleSelector couldn't be
unfocused. This was happening because if the field is "active" we pass
"focused=true" to the Search component, which calls "focus" on the input
each time the component updates. This allows us to re-focus the input
after selecting a profile.

The issue was:

- field was active, we would call focus()
- field was inactive, but focus was called after other field was
  activated
- field was active again