Consider inflations and fields before clearing normalizations
An example of the logic before this causing an issue:

TeamMember optionally can have "team" inflated

If we fetch TeamMembers with team inflated, it gets stored in the cache,
the UI displays something related to the team.

If we go to another page that fetches TeamMembers but doesn't inflate
the team, we were deleting the team from the TeamMember normalization.
When you went back to a page, it pulled the team member from the cache
and attempted to access the team normalization, but it had been deleted.

Now we won't delete the normalizations if we can detect that the item
was excluded from the incoming response.